+++ src/actions/simulate/simulations.js
@@ -16,7 +16,7 @@ var simulations = {
         Applies any set deceleration and acceleration to existing velocity
     */
     velocity: function (value, duration) {
-        value.velocity = value.velocity - speedPerFrame(value.deceleration, duration) + speedPerFrame(value.acceleration, duration);
+        value.velocity = value.velocity + speedPerFrame(value.acceleration, duration);
 
         return simulations.friction(value, duration);
     },