+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Popmotion adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.3.1] 2015-12-24
+
+### Removed
+- `deceleration` from `Simulate`.
+
+### Fixed
+- Fixed bug where `Tween` would start as ended if initalised during `update` ie via a `Sequence`.
+
 ## [4.3.0] 2015-12-17
 
 ### Added